Make tuples be subtypes of empty interfaces (like Any).
After this commit, a tuple is considered a subtype of `Any ${cap}`
if all of the elements of the tuple are subcaps of the cap.

This is a principle of least surprise issue, as it is surprising
that a type param constraint of `Any #any` does not behave the same
as an unconstrained type param, with the latter accepting tuple
type args, and the latter not doing so.

Previously, there was no way to allow tuples of any cardinality
as a type argument, while restricting what caps were allowed to
be in those tuples (such as for the concerns of aliasing or
sendability). For example, `class List[A]` in the `collections`
package allowed tuples as type args, but `class List[A: Any #share]`
in the `collections/persistent` package did not. Now they both
allow tuple type args, and the `Any #share` constrains the type arg
so that all tuple elements must be shareable (`val` or `tag`).

Resolves #1767.